Adapter hecho pero no aplicado
- Gente comentad el codigo porfavor 😭

+++ ProjectePatrons/Items/Definition/AugmentaPunts.java
@@ -1,5 +1,5 @@
 package Definition;
 
 public interface AugmentaPunts {
-	int SumaPunts();
+	double SumaPunts();
 }
