+++ ProjectePatrons/Items/Definition/TreuPunts.java
@@ -1,5 +1,5 @@
 package Definition;
 
 public interface TreuPunts {
-	int RestaPunts();
+	double RestaPunts();
 }
